resolucao exercicios 1012 e 1013

+++ uri1012.js
@@ -1,7 +1,7 @@
 var input = require('fs').readFileSync('/dev/stdin', 'utf8');
 var lines = input.split('\n');
 
-var raio = parseInt(lines.shift());
+var raio = parseFloat(lines.shift());
 const pi = parseFloat(3.14159);
 
 var volume = ((4.0 / 3) * pi * (raio ** 3));
